Edited some nav bar colors Reorganized pages
