+++ lib/pages/welcome.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class Welcome extends StatefulWidget {
+  @override
+  _WelcomeState createState() => _WelcomeState();
+}
+
+class _WelcomeState extends State<Welcome> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
